fix: fix subscription unit test
This commit contains fix of subscription unit test workflow
